refactor: individual shlagemon files and dynamic imports
