+++ src/data/shlagemons/mewteub.ts
@@ -1,0 +1,14 @@
+import type { BaseShlagemon } from '~/type'
+import { shlagemonTypes } from '../shlagemons-type'
+
+export const mewteub: BaseShlagemon = {
+  id: 'mewteub',
+  name: 'Mewteub',
+  color: '#f9e743',
+  description: `Mewteub est un né d’un bug génétique entre une prise HDMI et une entité cosmique sous anxiolytiques. On murmure qu’il aurait été cloné à partir d’un vieux câble péritel tombé dans une bassine de Monster Energy. Il flotte constamment, mais sans élégance — plutôt comme une chaussette dans un jacuzzi. Son corps translucide laisse entrevoir un réseau veineux qui clignote au rythme d’un modem 56k. Il émet des sons télépathiques, tous plus gênants les uns que les autres : soupirs, "hmm" douteux et parfois des extraits de discussions MSN de 2003. Il communique principalement par vibration de teub psychique, ce qui perturbe fortement les autres Shlagémons — et les enfants non accompagnés. Sa capacité spéciale, *Onde Malaise*, fige l’adversaire dans une gêne abyssale pendant plusieurs tours. On dit que croiser son regard provoque des flashbacks de soirées gênantes en boîte de province.
+Mewteub est rare. Trop rare. Et franchement, c’est peut-être mieux ainsi.`,
+  types: [shlagemonTypes.psy],
+  coefficient: 10,
+}
+
+export default mewteub
