+++ src/data/zones.ts
@@ -1,10 +1,24 @@
 import type { BaseShlagemon } from '~/type'
 import type { FightZoneId, Zone, ZoneType } from '~/type/zone'
-import { abraquemar, alaouakbar, bulgrosboule, canarchichon, carapouffe, dartagnan, goubite, mewteub, nosferachid, pikachiant, ptitocard, qulbudrogue, rouxPasCool, sacdepates, salamiches } from './shlagemons'
+import { abraquemar } from './shlagemons/abraquemar'
+import { alaouakbar } from './shlagemons/alaouakbar'
+import { bulgrosboule } from './shlagemons/bulgrosboule'
+import { canarchichon } from './shlagemons/canarchichon'
+import { carapouffe } from './shlagemons/carapouffe'
+import { dartagnan } from './shlagemons/dartagnan'
 import { emboli } from './shlagemons/emboli'
+import { goubite } from './shlagemons/goubite'
 import { metamorve } from './shlagemons/metamorve'
+import { mewteub } from './shlagemons/mewteub'
 import { nanmeouesh } from './shlagemons/nameouesh'
+import { nosferachid } from './shlagemons/nosferachid'
+import { pikachiant } from './shlagemons/pikachiant'
+import { ptitocard } from './shlagemons/ptitocard'
+import { qulbudrogue } from './shlagemons/qulbudrogue'
 import { ricardnin } from './shlagemons/ricardnin'
+import { rouxPasCool } from './shlagemons/rouxPasCool'
+import { sacdepates } from './shlagemons/sacdepates'
+import { salamiches } from './shlagemons/salamiches'
 import { sperectum } from './shlagemons/sperectum'
 
 interface ZoneDescription {
